Fix the vulkan "Clear" renderpass framebuffer use
The EFB framebuffer was always created using the renderpass with the
"load" pixelOp, but the PowerVR driver (incorrectly) seems to require
the framebuffer renderpass' loadOp to match the draw renderpass.

Instead, this creates a second framebuffer object to be used by the
render passes that are doing a clear.

This fixes corruption on PowerVR drivers in Wind Waker's intro, as the
clear would get dropped and stale framebuffer contents would remain.